Feat(streams): 분리된 레포지토리 병합, 타입 수정, 완성

+++ src/main/java/app/domain/forecast/model/dto/request/RealDataItem.java
@@ -29,7 +29,7 @@ public class RealDataItem {
 
 	private int realOrderQuantity;
 
-	private int realSalesRevenue;
+	private Long realSalesRevenue;
 
 	private int dayOfWeek;
